requirement.txt
streamlit
python-docx
scikit-learn
numpy # scikit-learn dependency
scipy # scikit-learn dependency